adding code

+++ core/src/data/value/binary_op/mod.rs
@@ -6,10 +6,10 @@ mod i128;
 mod i32;
 mod i64;
 mod i8;
-mod u8;
+mod u128;
 mod u32;
 mod u64;
-mod u128;
+mod u8;
 
 pub trait TryBinaryOperator {
     type Rhs;
